+++ extension/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## 1.12.2
+
+-   Added: Image to the readme to see how the extension looks
+-   Added: Autodetect registry type in "add user registry"
+-   Fixed: Some documentation was still NPM-only.
+
+## 1.12.1
+
+-   Added: Limit the number of visible extensions in the extension list
+-   Added: Search box to filter extensions
+
+## 1.12.0
+
+-   Updated: Mimic the look-and-feel of the normal vscode extension list
+
 ## 1.11.4
 
 -   Fixed: Fix parsing of vsix files in filesystem provider where nls substitution was used. Fixes #11